mv command gen dao to separate package gendao
